gitlab: move BoolValue type and associated methods to gitlab.go
